tools/scylla-sstable: extract json_writer into its own header
Other source files will want to use it soon.
