+++ tools/scylla-sstable.cc
@@ -30,15 +30,14 @@
 #include "types/user.hh"
 #include "types/set.hh"
 #include "types/map.hh"
+#include "tools/json_writer.hh"
 #include "tools/schema_loader.hh"
 #include "tools/utils.hh"
-#include "utils/rjson.hh"
 #include "locator/host_id.hh"
 
-// has to be below the utils/rjson.hh include
-#include <rapidjson/ostreamwrapper.h>
-
 using namespace seastar;
+
+using json_writer = tools::json_writer;
 
 namespace bpo = boost::program_options;
 
@@ -200,73 +199,6 @@ public:
             return make_ready_future<stop_iteration>(stop_iteration::yes);
         }
         return std::move(mf).consume(_consumer);
-    }
-};
-
-class json_writer {
-    using stream = rapidjson::BasicOStreamWrapper<std::ostream>;
-    using writer = rapidjson::Writer<stream, rjson::encoding, rjson::encoding, rjson::allocator>;
-
-    stream _stream;
-    writer _writer;
-
-public:
-    json_writer() : _stream(std::cout), _writer(_stream)
-    { }
-
-    // following the rapidjson method names here
-    bool Null() { return _writer.Null(); }
-    bool Bool(bool b) { return _writer.Bool(b); }
-    bool Int(int i) { return _writer.Int(i); }
-    bool Uint(unsigned i) { return _writer.Uint(i); }
-    bool Int64(int64_t i) { return _writer.Int64(i); }
-    bool Uint64(uint64_t i) { return _writer.Uint64(i); }
-    bool Double(double d) { return _writer.Double(d); }
-    bool RawNumber(std::string_view str) { return _writer.RawNumber(str.data(), str.size(), false); }
-    bool String(std::string_view str) { return _writer.String(str.data(), str.size(), false); }
-    bool StartObject() { return _writer.StartObject(); }
-    bool Key(std::string_view str) { return _writer.Key(str.data(), str.size(), false); }
-    bool EndObject(rapidjson::SizeType memberCount = 0) { return _writer.EndObject(memberCount); }
-    bool StartArray() { return _writer.StartArray(); }
-    bool EndArray(rapidjson::SizeType elementCount = 0) { return _writer.EndArray(elementCount); }
-
-    // scylla-specific extensions (still following rapidjson naming scheme for consistency)
-    template <typename T>
-    void AsString(const T& obj) {
-        String(fmt::format("{}", obj));
-    }
-    // partition or clustering key
-    template <typename KeyType>
-    void DataKey(const schema& schema, const KeyType& key, std::optional<dht::token> token = {}) {
-        StartObject();
-        if (token) {
-            Key("token");
-            AsString(*token);
-        }
-        Key("raw");
-        String(to_hex(key.representation()));
-        Key("value");
-        AsString(key.with_schema(schema));
-        EndObject();
-    }
-    void StartStream() {
-        StartObject();
-        Key("sstables");
-        StartObject();
-    }
-    void EndStream() {
-        EndObject();
-        EndObject();
-    }
-    void SstableKey(const sstables::sstable& sst) {
-        Key(sst.get_filename());
-    }
-    void SstableKey(const sstables::sstable* const sst) {
-        if (sst) {
-            SstableKey(*sst);
-        } else {
-            Key("anonymous");
-        }
     }
 };
